Created menu.jsx file
