+++ client/src/components/Avatar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-export const Avatar = ({ src }) => {
+export const Avatar = ({ src, size }) => {
   const { theme } = useSelector((state) => state);
 
   return (
@@ -9,7 +9,7 @@ export const Avatar = ({ src }) => {
       <img
         src={src}
         alt='avatar'
-        className='avatar'
+        className={size}
         style={{ filter: `${theme ? 'invert(1)' : 'invert(0)'}` }}
       />
     </>
