+++ client/src/index.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import './index.css';
+import './styles/global.css';
+
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { store } from './redux/store';
