+++ client/src/styles/avatar.css
@@ -1,0 +1,17 @@
+.big-avatar {
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+}
+
+.medium-avatar {
+  width: 30px;
+  height: 30px;
+  border-radius: 50%;
+}
+
+.small-avatar {
+  width: 20px;
+  height: 20px;
+  border-radius: 50%;
+}
