+++ client/src/styles/global.css
@@ -38,17 +38,14 @@ img {
   object-fit: cover;
 }
 
-.avatar {
-  width: 30px;
-  height: 30px;
-  border-radius: 50%;
-}
-
 /*------------------AUTH------------------*/
-@import url('./styles/auth.css');
+@import url('./auth.css');
 
 /*------------------LOADING------------------*/
-@import url('./styles/loading.css');
+@import url('./loading.css');
 
 /*------------------LOADING------------------*/
-@import url('./styles/header.css');
+@import url('./header.css');
+
+/*------------------AVATAR------------------*/
+@import url('./avatar.css');
